Extract SystemdStrategy
